feat(metrics): refactor SrtEndpointMetrics import path and update related references
